requirements.txt
djangorestframework_simplejwt==5.0.0